// apps/web/src/data/generated/ponder/clean-graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };

/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BigInt: { input: string; output: string; }
};

export enum OrderDirection {
  Asc = 'asc',
  Desc = 'desc'
}
// ...
export type VrgdaPoolSeedWhereInput = {
  blockNumber?: InputMaybe<Scalars['BigInt']['input']>;
  blockNumber_gt?: InputMaybe<Scalars['BigInt']['input']>;
  blockNumber_gte?: InputMaybe<Scalars['BigInt']['input']>;
  blockNumber_lt?: InputMaybe<Scalars['BigInt']['input']>;
  blockNumber_lte?: InputMaybe<Scalars['BigInt']['input']>;
  blockNumber_in?: InputMaybe<Array<Scalars['BigInt']['input']>>;
  isUsed?: InputMaybe<Scalars['Boolean']['input']>;
  background?: InputMaybe<Scalars['Int']['input']>;
  background_in?: InputMaybe<Array<Scalars['Int']['input']>>;
  body?: InputMaybe<Scalars['Int']['input']>;
  body_in?: InputMaybe<Array<Scalars['Int']['input']>>;
  accessory?: InputMaybe<Scalars['Int']['input']>;
  accessory_in?: InputMaybe<Array<Scalars['Int']['input']>>;
  head?: InputMaybe<Scalars['Int']['input']>;
  head_in?: InputMaybe<Array<Scalars['Int']['input']>>;
  glasses?: InputMaybe<Scalars['Int']['input']>;
  glasses_in?: InputMaybe<Array<Scalars['Int']['input']>>;
};